backend/requirements.txt
# Core
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.7,<3
pydantic-settings>=2.3,<3

# Database
sqlalchemy[asyncio]>=2.0,<3
aioodbc>=0.5,<1
pyodbc>=5,<6
alembic>=1.13,<2

# Redis & Queue
redis>=5.0,<6
celery[redis]>=5.4,<6

# AI / Agent
langchain>=0.3,<1
langchain-openai>=0.2,<1
langchain-community>=0.3,<1
chromadb>=0.5,<1
tiktoken>=0.7,<1

# Auth & Security
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
cryptography>=42,<44

# Observability
structlog>=24,<25
opentelemetry-api>=1.21,<2
opentelemetry-sdk>=1.21,<2
opentelemetry-instrumentation-fastapi>=0.42,<1

# HTTP
httpx>=0.27,<1
python-multipart>=0.0.9,<1

# Document parsing
pypdf>=4,<5

# Dev
ruff>=0.5,<1
mypy>=1.10,<2
pytest>=8.2,<9
pytest-asyncio>=0.23,<1
